restructure of test files, test.bash upgrade
