+++ servers/lib/script/test.bash
@@ -3,11 +3,32 @@ echo "testing in progress"
 new_path="$(yarn bin):$PATH"
 export PATH="$new_path"
 
-echo "making unit tests"
-jest ../test
+# Get the first argument passed to the script
+mode=$1
 
-echo "making e2e tests"
-jest --config ./test/jest-e2e.json
+if [ "$mode" == "-a" ]; then
+  echo "making all tests ..."
 
-echo "retrieving coverage report"
-jest --coverage
+  echo "making unit tests"
+  jest ../test/unit
+
+  echo "making integration tests"
+  jest ../test/integration 
+
+  echo "making e2e tests"
+  jest --config ./test/jest-e2e.json
+  
+  echo "retrieving coverage report"
+  jest --coverage
+elif [ "$mode" == "-u" ]; then
+  echo "making unit tests"
+  jest ../test/unit
+elif [ "$mode" == "-i" ]; then
+  echo "making integration tests"
+  jest ../test/integration
+elif [ "$mode" == "-e" ]; then
+  echo "making e2e tests"
+  jest --config ./test/jest-e2e.json
+else
+  echo "Invalid option. Please use -a for all tests, -u for unit tests, -i for integration tests, or -e for e2e tests."
+fi
