+++ servers/lib/test/e2e/app.e2e.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from "@nestjs/testing";
 import { INestApplication } from "@nestjs/common";
 import * as request from "supertest";
 import { execSync } from "child_process";
-import { AppModule } from "../src/app.module";
+import { AppModule } from "../../src/app.module";
 
 describe("End to End test for the application", () => {
   let app: INestApplication;
@@ -16,7 +16,7 @@ describe("End to End test for the application", () => {
     await app.listen(process.env.PORT);
   });
 
-  afterEach(async () => {
+  afterAll(async () => {
     await app.close();
   });
