+++ servers/lib/test/unit/files.resolver.unit.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
-import { FilesService } from "../src/files/files.service";
-import { FilesResolver } from "../src/files/files.resolver";
+import { FilesService } from "../../src/files/files.service";
+import { FilesResolver } from "../../src/files/files.resolver";
 
 describe("Unit tests for FilesResolver", () => {
   let filesResolver: FilesResolver;
